site_analytics/requirements.txt
Django==1.6.2
Werkzeug==0.9.4
django-debug-toolbar==1.0.1
django-extensions==1.3.3
django-jsonfield==0.9.12
djangorestframework==2.3.13
factory-boy==2.3.1
pytz==2014.1.1
six==1.6.1
sqlparse==0.1.11
wsgiref==0.1.2
